Modified the NEB to use RDKit estimates
We would optimize the RDKit estimate in the
qmcalculator (MOPAC, G09, etc), then use
the optimized geometries to start double-ended
searches. That strategy may help for QST2 calcs
but creates problems for the NEB. The optimized
product may not result from the reactant conformer,
causing problems along the elastic band.

The original strategy, where we create the product geometry
by optimizing the reactant with the product bounds matrix,
is more likely to create a product conformer that would be made from
the reactant. So reverting to this should avoid errors in the NEB
calculations (and reduce core dumps).